ProjetGroupe/projectblog/requirements.txt
anyjson==0.3.3
Django==2.2.7
django-admin-generator==2.2.0
django-filebrowser-no-grappelli==3.7.8
django-tinymce4-lite==1.7.5
django-utils2==2.5.4
djangorestframework==3.10.3
jsmin==2.2.2
Pillow==6.2.1
python-utils==2.3.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
psycopg2-binary
django_filters
drf-dynamic-fields
